+++ include/swift/Sema/PreparedOverload.h
@@ -27,6 +27,34 @@ namespace constraints {
 class ConstraintLocatorBuilder;
 class ConstraintSystem;
 
+/// Describes the type produced when referencing a declaration.
+struct DeclReferenceType {
+  /// The "opened" type, which is the type of the declaration where any
+  /// generic parameters have been replaced with type variables.
+  ///
+  /// The mapping from generic parameters to type variables will have been
+  /// recorded by \c recordOpenedTypes when this type is produced.
+  Type openedType;
+
+  /// The opened type, after performing contextual type adjustments such as
+  /// removing concurrency-related annotations for a `@preconcurrency`
+  /// operation.
+  Type adjustedOpenedType;
+
+  /// The type of the reference, based on the original opened type. This is the
+  /// type that the expression used to form the declaration reference would
+  /// have if no adjustments had been applied.
+  Type referenceType;
+
+  /// The type of the reference, which is the adjusted opened type after
+  /// (e.g.) applying the base of a member access. This is the type of the
+  /// expression used to form the declaration reference.
+  Type adjustedReferenceType;
+
+  /// The type that could be thrown by accessing this declaration.
+  Type thrownErrorTypeOnAccess;
+};
+
 /// Describes a dependent type that has been opened to a particular type
 /// variable.
 using OpenedType = std::pair<GenericTypeParamType *, TypeVariableType *>;
@@ -101,6 +129,7 @@ struct PreparedOverload {
   };
 
   ArrayRef<Change> Changes;
+  DeclReferenceType DeclType;
 };
 
 struct PreparedOverloadBuilder {
